Move get test from project into individual method.
